+++ e-farma/app/controllers/PreservesController.scala
@@ -1,5 +1,6 @@
 package controllers
 
+import controllers.request.Common
 import javax.inject._
 import models.entities.Preserves
 import models.forms.{PreservesForm, PreservesUpdateFormData}
@@ -11,12 +12,20 @@ import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 @Singleton
-class PreservesController @Inject()(cc: MessagesControllerComponents, preservesService: PreservesService) extends MessagesAbstractController(cc) {
+class PreservesController @Inject()(cc: MessagesControllerComponents, preservesService: PreservesService, common: Common) extends MessagesAbstractController(cc) {
 
   def getAll(): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    preservesService.listAllItems map ( items =>
-      Ok(Json.toJson(items))
-      )
+    if (common.checkAuth(request.headers.toMap("Cookie").toString())) {
+      preservesService.listAllItems map (items =>
+        Ok(Json.toJson(items))
+        )
+    }
+    else {
+
+      Future {
+        Forbidden("Wrong Auth")
+      }
+    }
   }
 
   def getAllPreservesView(): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
@@ -25,10 +34,19 @@ class PreservesController @Inject()(cc: MessagesControllerComponents, preservesS
   }
 
   def getById(id: Long): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    preservesService.getItem(id) map { item =>
-      Ok(Json.toJson(item))
+    if (common.checkAuth(request.headers.toMap("Cookie").toString())) {
+      preservesService.getItem(id) map { item =>
+        Ok(Json.toJson(item))
+      }
+    }
+    else {
+
+      Future {
+        Forbidden("Wrong Auth")
+      }
     }
   }
+
   def getByIdPreserveView(id: Long): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
     val preserve = preservesService.getItem(id)
     preserve.map(honey => honey match {
@@ -39,15 +57,23 @@ class PreservesController @Inject()(cc: MessagesControllerComponents, preservesS
 
 
   def add(): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    PreservesForm.form.bindFromRequest().fold(
-      errorForm => {
-        errorForm.errors.foreach(println)
-        Future.successful(BadRequest("Error!"))
-      },
-      data => {
-        val newSeedItem = Preserves(0, data.name, data.quantity, data.weight, data.price, data.description)
-        preservesService.addItem(newSeedItem).map(_ => Redirect(routes.PreservesController.getAll()))
-      })
+    if (common.checkAuth(request.headers.toMap("Cookie").toString())) {
+      PreservesForm.form.bindFromRequest().fold(
+        errorForm => {
+          errorForm.errors.foreach(println)
+          Future.successful(BadRequest("Error!"))
+        },
+        data => {
+          val newSeedItem = Preserves(0, data.name, data.quantity, data.weight, data.price, data.description)
+          preservesService.addItem(newSeedItem).map(_ => Redirect(routes.PreservesController.getAll()))
+        })
+    }
+    else {
+
+      Future {
+        Forbidden("Wrong Auth")
+      }
+    }
   }
 
   def addPreserveView: Action[AnyContent] = Action.async { implicit request =>
@@ -68,15 +94,23 @@ class PreservesController @Inject()(cc: MessagesControllerComponents, preservesS
   }
 
   def updatePreserve: Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    PreservesForm.updateForm.bindFromRequest().fold(
-      errorForm => {
-        errorForm.errors.foreach(println)
-        Future.successful(BadRequest("Error!"))
-      },
-      data => {
-        val seedItem = Preserves(data.id, data.name, data.quantity, data.weight, data.price, data.description)
-        preservesService.updateItem(seedItem).map(_ => Redirect(routes.PreservesController.getAll()))
-      })
+    if (common.checkAuth(request.headers.toMap("Cookie").toString())) {
+      PreservesForm.updateForm.bindFromRequest().fold(
+        errorForm => {
+          errorForm.errors.foreach(println)
+          Future.successful(BadRequest("Error!"))
+        },
+        data => {
+          val seedItem = Preserves(data.id, data.name, data.quantity, data.weight, data.price, data.description)
+          preservesService.updateItem(seedItem).map(_ => Redirect(routes.PreservesController.getAll()))
+        })
+    }
+    else {
+
+      Future {
+        Forbidden("Wrong Auth")
+      }
+    }
   }
 
   def updatePreserveView(id: Long): Action[AnyContent] = Action.async { implicit request: MessagesRequest[AnyContent] =>
@@ -88,8 +122,16 @@ class PreservesController @Inject()(cc: MessagesControllerComponents, preservesS
   }
 
   def delete(id: Long): Action[AnyContent] = Action.async { implicit request: Request[AnyContent] =>
-    preservesService.deleteItem(id) map { res =>
-      Redirect(routes.PreservesController.getAll())
+    if (common.checkAuth(request.headers.toMap("Cookie").toString())) {
+      preservesService.deleteItem(id) map { res =>
+        Redirect(routes.PreservesController.getAll())
+      }
+    }
+    else {
+
+      Future {
+        Forbidden("Wrong Auth")
+      }
     }
   }
 }
